Moved playing sounds to SoundService.tsx
